refactor: thunk 메서드를 분리하고, useCart를 리팩토링한다
