+++ src/components/@mixins/ResultBox/ResultBox.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { formatPrice } from "../../../utils/utils";
+import { formatPrice } from "../../../utils/util";
 import Button from "../../@shared/Button/Button";
 import * as S from "./ResultBox.styled";
