+++ src/hooks/useCart.js
@@ -1,33 +1,55 @@
 import { useDispatch, useSelector } from "react-redux";
 import {
-  addToCart,
-  getCarts as _getCarts,
-  removeChecked,
-  removeFromCart,
+  addToCartAsync,
+  getCartsAsync,
+  removeCheckedAsync,
+  removeFromCartAsync,
+} from "../store/modules/cart/cartThunk";
+import {
   toggleChecked,
   toggleAllChecked as _toggleAllChecked,
   resetError,
-} from "../store/modules/cartSlice";
+} from "../store/modules/cart/cartSlice";
 
 // eslint-disable-next-line import/prefer-default-export
 export const useCart = () => {
   const dispatch = useDispatch();
   const { items, loading, errorMessage } = useSelector((state) => state.cart);
   const carts = Object.values(items);
+  const cartAmount = carts.reduce((acc, cur) => acc + cur.amount, 0);
+  const checkedItems = carts.filter((item) => item.checked);
+  const hasCheckedItems = checkedItems.length > 0;
+  const totalPrice = checkedItems.reduce(
+    (acc, { amount, price }) => acc + amount * price,
+    0
+  );
+  const checkedSet = new Set(carts.map(({ checked }) => checked));
+  let isCheckAll = null;
+  switch (checkedSet.size) {
+    case 0:
+      isCheckAll = false;
+      break;
+    case 1:
+      isCheckAll = [...checkedSet].pop();
+      break;
+    default:
+  }
+
+  const getCartAmount = (id) => items[id]?.amount ?? 0;
 
   const getCarts = () => {
-    dispatch(_getCarts());
+    dispatch(getCartsAsync());
   };
 
   const addCart = (item) => {
-    dispatch(addToCart(item));
+    dispatch(addToCartAsync(item));
   };
 
   const removeCart = (item) => {
-    dispatch(removeFromCart(item));
+    dispatch(removeFromCartAsync(item));
   };
   const removeCheckedCart = (cart) => {
-    dispatch(removeChecked(cart));
+    dispatch(removeCheckedAsync(cart));
   };
 
   const toggleCartChecked = (id) => {
@@ -42,44 +64,22 @@ export const useCart = () => {
     dispatch(resetError());
   };
 
-  const getCartAmount = (id) => items[id]?.amount ?? 0;
-
-  const cartAmount = carts.reduce((acc, cur) => acc + cur.amount, 0);
-  const checkedItems = carts.filter((item) => item.checked);
-  const hasCheckedItems = checkedItems.length > 0;
-  const totalPrice = checkedItems.reduce(
-    (acc, { amount, price }) => acc + amount * price,
-    0
-  );
-
-  const checkedSet = new Set(carts.map(({ checked }) => checked));
-  let isCheckAll = null;
-  switch (checkedSet.size) {
-    case 0:
-      isCheckAll = false;
-      break;
-    case 1:
-      isCheckAll = [...checkedSet].pop();
-      break;
-    default:
-  }
-
   return {
     items,
     loading,
     errorMessage,
-    resetCartError,
+    cartAmount,
+    checkedItems,
+    hasCheckedItems,
+    isCheckAll,
+    totalPrice,
+    getCartAmount,
     getCarts,
     addCart,
     removeCart,
     removeCheckedCart,
     toggleCartChecked,
     toggleAllChecked,
-    getCartAmount,
-    cartAmount,
-    checkedItems,
-    hasCheckedItems,
-    isCheckAll,
-    totalPrice,
+    resetCartError,
   };
 };
